1. There is a new constraint added: In the EvidenceRef element, the IdentificationFile attribute and the id_refs attribute MUST be present at the same time Accordingly the codes are changed: xTracker.java, outputMZQ.java and study.java
2. Make it consistent that IDs from identification files keep unchanged until the output step: loadMascotIdent.java loadMzIdentML.java, xPeptide.java

3. update the example files due to point 1 and point 3 in revision 235 (in csv protein and peptide order)

4. found the maven entry for Jama: delete the java files and add the dependency in the pom file

5. Remove several commented out lines in various files

known bug:
As the current jmzQuantML only accept the actual object(e.g. Protein), not String, the outputMZQ plugin has to assign the id corrected for NCName to the object, which will affect the result in other output plugins if used after outputMZQ plugin. 
e.g. Protein object has the id sp|P04004, after the correction, the object has the id as sp_P04004. The other plugin will access the same object and output the id as sp_P04004, not the desired sp|P04004. 
The solution is to put other output plugins in front of outputMZQ in the pipeline or add extra field in xProtein to store the original id and all output plugins use the original id.