The commit includes the protein ratio from peptide ratio inference method from the MCP paper by Hultin-Rosenberg et al (variant based on the original method by Onsongo et al can also be used).
xTracker.java
Previously the method used for calculating protein ratios from peptide ratios was a median, and that now is set by default to the PROTEIN_PEPTIDE_RATIO_INFERENCE in xTracker.java. The new method can be set by setting a userParam "protein ratio from peptide ratio calculation method" to "intensityWeightedAverage" in the pipeline feature detection and quantitation step. Accordingly, "intensityWeightedAverage" is now an option in getInferenceMethod().

GenericInferenceMethod.java
Added method intensityWeightedAverage() used by the algorithm. It's made to be generic, accepting an arbitrary list of weights for the ratio values it receives.

Utils.java
Added a logarithm method for log2, used quite extensively by the inference method.

iTraqQuantitation.java
Refactored start() to quantitate peptides and proteins in separate loops. The function of the code remains unchanged, but now all the peptides are quantitated before any protein quantitation happens. This was necessary to make the ratio inference algorithm work. Any decrease in performance cause by more iterating over proteins/peptides should be negligible.

QuantitationLevel.java
Modified calculateRatioFromPeptideRatios() to accept differnt ratio inference methods. The bulk of the algorithms functionality is implemented here in the IntensityWeights class. It could be split cleanly into a separate file, although I wasn't sure which package would be most appropriate to put it in. The class is used as a static property of QuantitationLevel. Several parameters are currently hard coded for the test data I've used, to be changed into parameters soon (need to consult with Jun).